stickies extend stickie model issue EMPCP-86
git-svn-id: https://svn.intranet.ydp/ydp/projects/empiria/player/trunk/empiria.player@123587 3e83ccf1-c6ce-0310-9054-8fdb33cf0640

+++ src/eu/ydp/empiria/player/client/controller/extensions/internal/stickies/StickieProperties.java
@@ -69,11 +69,11 @@ public class StickieProperties extends JavaScriptObject implements IStickiePrope
 		this.timestamp = new Date().getTime();
 	}-*/;
 	
-	public final native void setTimestamp(long ts)/*-{
+	public final native void setTimestamp(double ts)/*-{
 		this.timestamp = ts;
 	}-*/;
 	
-	public final native long getTimestamp()/*-{
+	public final native double getTimestamp()/*-{
 		return this.timestamp;
 	}-*/;
